requirements.txt
streamlit
openai
faiss-cpu
numpy
requests
beautifulsoup4
python-dotenv
